+++ pyproject.toml
@@ -1,0 +1,22 @@
+[project]
+name = "fwi"
+version = "0.1.0"
+authors = [
+  { name="", email="" },
+]
+description = "Fire Weather Index (FWI)"
+readme = "README.md"
+requires-python = ">=3.6"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
+dependencies = [
+    "numpy",
+    "pandas",
+    "xarray",
+]
+
+[project.scripts]
+run_wildfires_fwi = "fwi.run_wildfires_fwi:cli"
